# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: API for bank cards, users and transfers management
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Returns information about authenticated user. JWT token requires
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto_UserResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token format: `Bearer <token>`
        Token may be retrieved `/api/v1/auth/login`
  schemas:
...
    Role:
      type: string
      enum: [USER, ADMIN]
      description: User role
      example: USER
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: User login
          example: rvkuzmich
        email:
          type: string
          format: email
          description: User Email
          example: r.kuzmich@example.com
        firstName:
          type: string
          description: User first name
          example: Roman
        lastName:
          type: string
          description: User last name
          example: Kuzmich
        role:
          $ref: '#/components/schemas/Role'
        enabled:
          type: boolean
          description: Is user account enabled
          example: true
        createdAt:
          type: string
          format: date-time
          description: User account creation timestamp
          example: '2024-01-15T10:30:00'
        cardCount:
          type: integer
          description: User cards count
          example: 3
...
    ApiResponseDto:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success flag
          example: true
        message:
          type: string
          description: Message text
          example: Operation completed successfully
        data:
          type: object
          description: Response data (may be null in case of error)
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2024-01-15T14:30:00'
    ErrorResponseDto:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: Operation success flag
          example: false
        message:
          type: string
          description: Main error message
          example: Validation failed
        errors:
          type: array
          description: Detailed errors message (may be null)
          items:
            type: string
          example: ["Email must be valid", "Password is required"]
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T14:30:00'
...
    ApiResponseDto_UserResponseDto:
      allOf:
        - $ref: '#/components/schemas/ApiResponseDto'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserResponseDto'
...